Move `Match` into its own module
