+++ src/lib.rs
@@ -13,13 +13,10 @@ pub use color_eyre;
 use color_eyre::eyre::{eyre, Result};
 use crossbeam_channel::{unbounded, Receiver, Sender};
 use dependencies::{Build, BuildManager};
-use lazy_static::lazy_static;
 use parser::{ErrorMatch, ErrorMatchKind, OptWithLine, Revisioned, Spanned};
-use regex::bytes::{Captures, Regex};
 use rustc_stderr::{Level, Message};
 use spanned::Span;
 use status_emitter::{StatusEmitter, TestStatus};
-use std::borrow::Cow;
 use std::collections::{HashSet, VecDeque};
 use std::ffi::OsString;
 use std::num::NonZeroUsize;
@@ -34,12 +31,14 @@ mod config;
 mod dependencies;
 mod diff;
 mod error;
+pub mod filter;
 pub mod github_actions;
 mod mode;
 mod parser;
 pub mod per_test_config;
 mod rustc_stderr;
 pub mod status_emitter;
+
 #[cfg(test)]
 mod tests;
 
@@ -49,69 +48,6 @@ pub use error::*;
 pub use mode::*;
 
 pub use spanned;
-
-/// A filter's match rule.
-#[derive(Clone, Debug)]
-pub enum Match {
-    /// If the regex matches, the filter applies
-    Regex(Regex),
-    /// If the exact byte sequence is found, the filter applies
-    Exact(Vec<u8>),
-    /// Uses a heuristic to find backslashes in windows style paths
-    PathBackslash,
-}
-impl Match {
-    fn replace_all<'a>(&self, text: &'a [u8], replacement: &[u8]) -> Cow<'a, [u8]> {
-        match self {
-            Match::Regex(regex) => regex.replace_all(text, replacement),
-            Match::Exact(needle) => text.replace(needle, replacement).into(),
-            Match::PathBackslash => {
-                lazy_static! {
-                    static ref PATH_RE: Regex = Regex::new(
-                        r"(?x)
-                        (?:
-                            # Match paths to files with extensions that don't include spaces
-                            \\(?:[\pL\pN.\-_']+[/\\])*[\pL\pN.\-_']+\.\pL+
-                        |
-                            # Allow spaces in absolute paths
-                            [A-Z]:\\(?:[\pL\pN.\-_'\ ]+[/\\])+
-                        )",
-                    )
-                    .unwrap();
-                }
-
-                PATH_RE.replace_all(text, |caps: &Captures<'_>| {
-                    caps[0].replace(r"\", replacement)
-                })
-            }
-        }
-    }
-}
-
-impl From<&'_ Path> for Match {
-    fn from(v: &Path) -> Self {
-        let mut v = v.display().to_string();
-        // Normalize away windows canonicalized paths.
-        if v.starts_with(r"\\?\") {
-            v.drain(0..4);
-        }
-        let mut v = v.into_bytes();
-        // Normalize paths on windows to use slashes instead of backslashes,
-        // So that paths are rendered the same on all systems.
-        for c in &mut v {
-            if *c == b'\\' {
-                *c = b'/';
-            }
-        }
-        Self::Exact(v)
-    }
-}
-
-impl From<Regex> for Match {
-    fn from(v: Regex) -> Self {
-        Self::Regex(v)
-    }
-}
 
 /// Run all tests as described in the config argument.
 /// Will additionally process command line arguments.
